pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["André Oliveira <andreaoliveira@ua.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.0"
uvicorn = "^0.31.1"
sqlalchemy = "^2.0.35"
pydantic = "^2.9.2"
python-dotenv = "^1.0.1"
boto3 = "^1.35.38"
pymysql = "^1.1.1"
requests = "^2.32.3"
cryptography = "^43.0.1"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
